Fixed data type for angles to be float (instead of int8_t in some cases). Also simplified cone calc.

+++ src/world/Action/ActionShapeLut.h
@@ -8,8 +8,8 @@
 namespace Sapphire::World::Action
 {
   struct ConeEntry {
-    int16_t startAngle;
-    int16_t endAngle;
+    float startAngle;
+    float endAngle;
   };
 
   class ActionShapeLut
